Don't copy the response Transfer-Encoding header
This stops `express-http-proxy` from preserving the response's
`Transfer-Encoding` header.

Without this, when `express-http-proxy` tries to proxy a response with
`Transfer-Encoding: chunked` set, the response is de-chunked by waiting
on all the chunks and joining them, indeed, a Content-Lenght header is
computed and set if intercepting. But trying to send a response with
both `Content-Length` and `Transfer-Encodign: chunked` fails with a
`Parse error`, because it doesn't make sense.

Thus, we explicitly skip copying over `Transfer-Encoding`.